Extract old card panel
